renaming userList to proposalList
