+++ src/store/modules/proposalList/actions.js
@@ -1,8 +1,10 @@
-import helpers from '../../../helpers';
+import Helpers from '../../../helpers';
+
+const helpers = new Helpers();
 
 export const gets = async () => {
   try {
-    const result = await helpers.authInstance.get('api/user/');
+    const result = await helpers.authInstance.get('api/user/proposal');
     return result.data;
   } catch (error) {
     return error.response.data;
