--- src/store/modules/userList/index.js
@@ -1,37 +1,0 @@
-import * as types from '../../../components/UserList/mutation-types';
-
-import { gets } from './actions';
-
-const initialState = {
-  users: [],
-};
-
-// getters
-const getters = {
-  users: state => state.users,
-};
-
-// actions
-const actions = {
-  async gets({ commit }, payload) {
-    const result = await gets(payload);
-    commit(types.GETS, result);
-  },
-};
-
-const mutations = {
-  GETS(state, payload) {
-    if (payload.meta.success) {
-      state.users = payload.data;
-    } else {
-      state.users = [];
-    }
-  },
-};
-
-export default {
-  state: initialState,
-  getters,
-  actions,
-  mutations,
-};
